Create csv::internals::speculative
