+++ include/internal/basic_csv_parser_guessing.cpp
@@ -1,6 +1,6 @@
 #include "basic_csv_parser.hpp"
-#include "csv_speculative_parser.hpp"
 #include "string_view_stream.hpp"
+#include "stream_parser.hpp"
 
 #include <unordered_map>
